refactor: split error handling utilities (#13)
