+++ tests/unit/test_error_handling_module_split.py
@@ -1,0 +1,25 @@
+"""Regression tests for the error_handling module split."""
+
+from open_stocks_mcp.tools import error_handling
+from open_stocks_mcp.tools.responses import create_success_response
+from open_stocks_mcp.tools.validation import validate_symbol
+
+
+def test_error_handling_reexports_core_symbols() -> None:
+    assert error_handling.execute_with_retry is not None
+    assert error_handling.classify_error is not None
+    assert error_handling.create_error_response is not None
+    assert error_handling.validate_symbol is not None
+
+
+def test_validation_behavior_preserved() -> None:
+    assert validate_symbol("AAPL") is True
+    assert validate_symbol("AAPL7") is True
+    assert validate_symbol("TOO-LONG") is False
+
+
+def test_success_response_default_status() -> None:
+    payload = {"symbol": "AAPL"}
+    response = create_success_response(payload)
+    assert response["result"]["status"] == "success"
+    assert response["result"]["symbol"] == "AAPL"
